requirements.txt
# Masumi payment network integration
masumi>=0.1.41
python-dotenv
uvicorn>=0.27.0

# Phoenix ML model dependencies (optional - install for recommendations)
jax==0.8.1
dm-haiku>=0.0.13
numpy>=1.26.4

# Additional utilities
# redis  # Uncomment for caching in production
# prometheus-client  # Uncomment for metrics monitoring
